Ninja Gold
For this assignment, use your knowledge of event handlers, angular components, and services to re-create the beloved Ninja Gold assignment using Angular, Node and MongoDB. The data should be stored in the database. For now, have a new game begin every time the page is refreshed - you may change this later if you choose.

BONUS: Create a leader board to show the top five scores.

BONUS: Allow a user to continue a saved game or start from scratch - not implemented

BONUS: Implement a save button so that the user may decide when the game is saved to the database.